implemented multi ops commands + minor improvements

+++ utils/clock.go
@@ -2,10 +2,10 @@ package utils
 
 import "time"
 
-func GetCurrentEPochTime() uint32 {
-	return uint32(time.Now().Unix())
+func GetCurrentEPochTime() int64 {
+	return time.Now().Unix()
 }
 
-func SecondsSince(since uint32) uint32 {
+func SecondsSince(since int64) int64 {
 	return GetCurrentEPochTime() - since
 }
